Implemented main audiobook creation functionality
Created the general layout and page structure. Implemented the audiobook
creation page, with OpenAI for the text to speech. Set it up as a POST
request. Added options for voice (automatically works in other
languages). Set up firebase for saving audio files, and MongoDB for the
audiobook database.

To do:
- Add Authentication
- Add homepage with created audiobooks
- Add an audiobook listening page
- Add thumbnail creation option